[no ci] WIP saving progress. Most changes are complete except accessing private fields of `Task`.
