+++ applications/rq_eval/src/lib.rs
@@ -144,7 +144,7 @@ fn run_single(iterations: usize) -> Result<(), &'static str> {
     let overhead = hpet_timing_overhead()?;
     let mut task = Task::new(
         None,
-        None,
+        task::InheritedStates::FromTask(&task::get_my_current_task().unwrap()),
         |_, _| loop { }, // dummy failure function
     )?;
     task.name = String::from("rq_eval_single_task_unrunnable");
