+++ kernel/unwind/src/lib.rs
@@ -941,8 +941,7 @@ fn cleanup_unwinding_context(unwinding_context_ptr: *mut UnwindingContext) -> ! 
     #[cfg(not(downtime_eval))]
     warn!("cleanup_unwinding_context(): invoking the task_cleanup_failure function for task {:?}", current_task);
     
-    (current_task.failure_cleanup_function)(
-        task::ExitableTaskRef::obtain_for_unwinder(current_task),
-        cause,
-    )
+    let (exitable_taskref, failure_cleanup_function) = 
+        task::ExitableTaskRef::obtain_for_unwinder(current_task);
+    failure_cleanup_function(exitable_taskref, cause)
 }
